refactor CLI parsing
