+++ src/cirrus/deploy.py
@@ -8,11 +8,21 @@ Essentially a thin wrapper for the deploy_plugins
 module and its constituent bits
 
 """
+import pluggage.registry
 from argparse import ArgumentParser
-from cirrus.deploy_plugins import bootstrap_parser, get_plugin
 from cirrus.logger import get_logger
+from cirrus.configuration import load_configuration
+
 
 LOGGER = get_logger()
+
+
+def get_plugin(plugin_name):
+    factory = pluggage.registry.get_factory(
+        'deploy',
+        load_modules=['cirrus.plugins.deployers']
+    )
+    return factory(plugin_name)
 
 
 def build_parser():
@@ -25,15 +35,30 @@ def build_parser():
     parser = ArgumentParser(
         description='git cirrus deploy command'
     )
-    subparsers = parser.add_subparsers(dest='command')
-    bootstrap_parser(subparsers)
-    opts = parser.parse_args()
+    parser.add_argument(
+        '--plugin', '-p', dest='plugin', default=None,
+        help="Name of deployment plugin to load"
+        )
+    opts, _ = parser.parse_known_args()
     return opts
 
 
 def main():
-    opts = build_parser()
-    plugin = get_plugin(opts.command)
+    initial_opts = build_parser()
+    pname = initial_opts.plugin
+    if initial_opts.plugin is None:
+        config = load_configuration()
+        if 'deploy' not in config:
+            msg = "No Plugin specified with --plugin or in cirrus.conf for deploy"
+            raise RuntimeError(msg)
+        pname = config.get_param('deploy', 'plugin', None)
+    if pname is None:
+        msg = "No Plugin specified with --plugin or in cirrus.conf for deploy"
+        raise RuntimeError(msg)
+
+    plugin = get_plugin(pname)
+    plugin.build_parser()
+    opts = plugin.parser.parse_args()
     plugin.deploy(opts)
