+++ src/cirrus/plugins/deployers/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+"""
+_deployers_
+
+Plugins for the deploy command
+
+"""
+import chef
